+++ packages/ice/src/plugins/web/index.ts
@@ -20,9 +20,9 @@ const plugin: Plugin = () => ({
   name: 'plugin-web',
   setup: ({ registerTask, onHook, context, generator, serverCompileTask, dataCache, watch, getAllPlugin }) => {
     const { rootDir, commandArgs, command, userConfig } = context;
-    const { ssg, dataLoader } = userConfig;
+    const { ssg } = userConfig;
 
-    registerTask(WEB, getWebTask({ rootDir, command, dataCache, dataLoader }));
+    registerTask(WEB, getWebTask({ rootDir, command, dataCache, userConfig }));
 
     generator.addExport({
       specifier: ['Link', 'Outlet', 'useParams', 'useSearchParams', 'useLocation', 'useNavigate'],