+++ packages/ice/src/service/analyze.ts
@@ -6,15 +6,13 @@ import moduleLexer from '@ice/bundles/compiled/es-module-lexer/index.js';
 import { esbuild } from '@ice/bundles';
 import type { Loader, Plugin } from 'esbuild';
 import consola from 'consola';
-import type { TaskConfig } from 'build-scripts';
-import type { Config } from '@ice/webpack-config/esm/types';
 import { getCache, setCache } from '../utils/persistentCache.js';
 import { getFileHash } from '../utils/hash.js';
 import scanPlugin from '../esbuild/scan.js';
 import type { DepScanData } from '../esbuild/scan.js';
 import formatPath from '../utils/formatPath.js';
 
-type Alias = TaskConfig<Config>['config']['alias'];
+type Alias = Record<string, string>;
 
 interface Options {
   parallel?: number;
@@ -162,22 +160,25 @@ interface ScanOptions {
   depImports?: Record<string, DepScanData>;
   plugins?: Plugin[];
   exclude?: string[];
+  ignores?: string[];
 }
 
 export async function scanImports(entries: string[], options?: ScanOptions) {
   const start = performance.now();
-  const { alias = {}, depImports = {}, exclude = [], rootDir, plugins } = options;
+  const { alias = {}, depImports = {}, exclude = [], rootDir, plugins, ignores } = options;
   const deps = { ...depImports };
 
   try {
     await Promise.all(
       entries.map((entry) =>
         esbuild.build({
+          alias,
           absWorkingDir: rootDir,
           write: false,
           entryPoints: [entry],
           bundle: true,
           format: 'esm',
+          platform: 'node',
           logLevel: 'silent',
           loader: { '.js': 'jsx' },
           plugins: [
@@ -185,6 +186,7 @@ export async function scanImports(entries: string[], options?: ScanOptions) {
               rootDir,
               deps,
               alias,
+              ignores,
               exclude,
             }),
             ...(plugins || []),