+++ packages/ice/src/tasks/web/index.ts
@@ -5,9 +5,16 @@ import { CACHE_DIR, RUNTIME_TMP_DIR } from '../../constant.js';
 import { getRoutePathsFromCache } from '../../utils/getRoutePaths.js';
 
 const require = createRequire(import.meta.url);
-const getWebTask = ({ rootDir, command, dataCache, dataLoader }): Config => {
+const getWebTask = ({ rootDir, command, dataCache, userConfig }): Config => {
   // basic task config of web task
   const defaultLogging = command === 'start' ? 'summary' : 'summary assets';
+  const removeExportExprs = ['serverDataLoader', 'staticDataLoader'];
+
+  // Remove dataLoader exports only when build in production
+  // and configure to generate data-loader.js.
+  if (command === 'build' && userConfig.dataLoader) {
+    removeExportExprs.push('dataLoader');
+  }
 
   return {
     mode: command === 'start' ? 'development' : 'production',
@@ -26,8 +33,8 @@ const getWebTask = ({ rootDir, command, dataCache, dataLoader }): Config => {
       ),
     },
     swcOptions: {
-      // When enable dataLoader, all dataLoader is built to data-loader.js.
-      removeExportExprs: dataLoader ? ['dataLoader', 'serverDataLoader', 'staticDataLoader'] : ['serverDataLoader', 'staticDataLoader'],
+      // The dataLoader is built by data-loader
+      removeExportExprs,
       keepPlatform: 'web',
       getRoutePaths: () => {
         return getRoutePathsFromCache(dataCache);