+++ packages/webpack-config/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.4
 
 - [feat] enable remove exports for all components
+- [fix] format id for win32 compatible
+- [fix] md4 -> md5 while md4 algorithm is not available anymore in NodeJS 17+ 
 
 ## 1.0.3